chore: Move enum derive to `enums::plain` in `avro_derive`
Prep for more complete enum support
